Move the single function of services.go into controller.go
